+++ benchmarks/go_to_c_adds/add_harmonic.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double add_harmonic(double total_so_far, int i) {
+double add_harmonic(double total_so_far, long i) {
   return total_so_far + 1.0 / i;
 }
