+++ benchmarks/go_to_c_adds/add_harmonic.h
@@ -1,6 +1,6 @@
 #ifndef ADD_HARMONIC_H
 #define ADD_HARMONIC_H
 
-double add_harmonic(double total_so_far, int i);
+double add_harmonic(double total_so_far, long i);
 
 #endif
