+++ benchmarks/go_to_c_concurrent/harmonic_range.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double add_harmonic(double total_so_far, int i) {
+double add_harmonic(double total_so_far, long i) {
   return total_so_far + 1.0 / i;
 }
 
-double harmonic_range(int start, int end) {
+double harmonic_range(long start, long end) {
   double sum = 0.0;
-  for (int i = start; i <= end; i++) {
+  for (long i = start; i <= end; i++) {
     sum = add_harmonic(sum, i);
   }
   return sum;
