+++ benchmarks/go_to_c_concurrent/harmonic_range.h
@@ -1,6 +1,6 @@
 #ifndef HARMONIC_RANGE_H
 #define HARMONIC_RANGE_H
 
-double harmonic_range(int start, int end);
+double harmonic_range(long start, long end);
 
 #endif
